Update README.md
Simplified hotkeys down a bit. Also, HotKeyControl does not like Shift in it's hotkey's very much since it is used while typing for capping letters.. This only seems to be a problem for certain combinations, like Ctrl+Shift+V because Ctrl+V already has a bind and the Shift doesn't differentiate it enough. Also, Tab has obvious problems because it tabs us out of the control we are trying to put a bind into.